pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ctf-bot"
version = "0.1.0"
description = "Discord bot for CTFtime and KCTF information"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10,<4.0",
    "beautifulsoup4>=4.12,<5.0",
    "discord.py>=2.4,<3.0",
    "python-dotenv>=1.0,<2.0",
]

[tool.setuptools]
packages = ["ctf_bot"]
